pyproject.toml
[tool.poetry]
name = "nomad"
version = "0.1.0"
description = ""
authors = ["DobbyDoesntFree <rkskekfkrk1@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.2"
beautifulsoup4 = "^4.11.1"
black = "^22.10.0"
pillow = "^9.3.0"
djangorestframework = "^3.14.0"


[tool.poetry.group.dev.dependencies]
pylint = "^2.15.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
